[REFACTOR] change from Converter to Assembler
