+++ src/main/java/techcourse/myblog/service/ArticleService.java
@@ -3,7 +3,7 @@ package techcourse.myblog.service;
 import org.springframework.stereotype.Service;
 import techcourse.myblog.domain.Article;
 import techcourse.myblog.domain.ArticleRepository;
-import techcourse.myblog.service.converter.ArticleConverter;
+import techcourse.myblog.service.assembler.ArticleAssembler;
 import techcourse.myblog.service.dto.ArticleDto;
 import techcourse.myblog.service.exception.NoArticleException;
 
@@ -13,19 +13,19 @@ import java.util.List;
 @Service
 public class ArticleService {
     private ArticleRepository articleRepository;
-    private final ArticleConverter articleConverter;
+    private final ArticleAssembler articleAssembler;
 
     public ArticleService(ArticleRepository articleRepository) {
         this.articleRepository = articleRepository;
-        this.articleConverter = ArticleConverter.getInstance();
+        this.articleAssembler = ArticleAssembler.getInstance();
     }
 
     public List<ArticleDto> findAll() {
-        return articleConverter.convertToEntities(articleRepository.findAll());
+        return articleAssembler.convertToEntities(articleRepository.findAll());
     }
 
     public Long post(ArticleDto articleDto) {
-        Article article = articleConverter.convertToEntity(articleDto);
+        Article article = articleAssembler.convertToEntity(articleDto);
         Article savedArticle = articleRepository.save(article);
         return savedArticle.getId();
     }
@@ -33,14 +33,14 @@ public class ArticleService {
     public ArticleDto findById(long articleId) {
         Article article = articleRepository.findById(articleId)
                 .orElseThrow(() -> new NoArticleException("해당 게시물은 존재하지 않습니다!"));
-        return articleConverter.convertToDto(article);
+        return articleAssembler.convertToDto(article);
     }
 
     @Transactional
     public void editArticle(ArticleDto articleDto, long articleId) {
         Article article = articleRepository.findById(articleId)
                 .orElseThrow(() -> new NoArticleException("해당 게시물은 존재하지 않습니다!"));
-        article.updateArticle(articleConverter.convertToEntity(articleDto));
+        article.updateArticle(articleAssembler.convertToEntity(articleDto));
     }
 
     public void deleteById(long articleId) {
